requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.7

# HTTP 客户端
aiohttp==3.9.1
requests==2.31.0

# 异步支持
asyncio==3.4.3

# 环境变量管理
python-dotenv==1.0.0

# 数据库
aiosqlite==0.19.0

# 日志
loguru==0.7.2

# 时间处理
python-dateutil==2.8.2

# JSON 处理
orjson==3.9.10

# 错误处理和重试
tenacity==8.2.3

# 类型提示
typing-extensions==4.9.0